Move NativeMethods in core and minor fixes
